requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
pytesseract==0.3.10
groq
python-dotenv
